arm/isis-obc/mci: Set RXRDY and TXRDY flags in PDC operations
This allows combining PDC with non-PDC read/write operations.  The
specification does not clearly indicate if RXRDY/TXRDY flags are being
set on PDC operations, however, according to the spec, the PDC reads
from the TDR/RDR registers and acts according to the flags. Which would
indicate that the flags are not masked if the PDCMODE bit is set.